+++ backend/tracim_backend/views/core_api/user_controller.py
@@ -238,6 +238,7 @@ class UserController(Controller):
             lang=hapic_data.body.lang,
             do_save=True
         )
+        uapi.execute_updated_user_actions(user)
         return uapi.get_user_with_context(user)
 
     @hapic.with_api_doc(tags=[SWAGGER_TAG__USER_ENDPOINTS])