+++ build_full_frontend.sh
@@ -90,9 +90,9 @@ cd $DEFAULTDIR/frontend_app_folder_advanced || exit 1
 
 
 # app Calendar
-log "cd $DEFAULTDIR/frontend_app_calendar"
-cd $DEFAULTDIR/frontend_app_calendar || exit 1
-./build_calendar.sh $dev
+log "cd $DEFAULTDIR/frontend_app_agenda"
+cd $DEFAULTDIR/frontend_app_agenda || exit 1
+./build_agenda.sh $dev
 
 # build Tracimtracim
 log "cd $DEFAULTDIR/frontend"