+++ frontend/src/container/Tracim.jsx
@@ -30,7 +30,8 @@ import {
   getAppList,
   getContentTypeList,
   getUserIsConnected,
-  getMyselfWorkspaceList
+  getMyselfWorkspaceList,
+  putUserLang
 } from '../action-creator.async.js'
 import {
   newFlashMessage,
@@ -83,12 +84,17 @@ class Tracim extends React.Component {
     const fetchGetUserIsConnected = await props.dispatch(getUserIsConnected())
     switch (fetchGetUserIsConnected.status) {
       case 200:
+        if (fetchGetUserIsConnected.json.lang === null) this.setDefaultUserLang(fetchGetUserIsConnected.json)
+
         props.dispatch(setUserConnected({
           ...fetchGetUserIsConnected.json,
           logged: true
         }))
+
         Cookies.set('lastConnection', '1', {expires: 180})
+
         i18n.changeLanguage(fetchGetUserIsConnected.json.lang)
+
         this.loadAppConfig()
         this.loadWorkspaceList()
         break
@@ -117,7 +123,7 @@ class Tracim extends React.Component {
   loadWorkspaceList = async (idOpenInSidebar = undefined) => {
     const { props } = this
 
-    const idWsToOpen = idOpenInSidebar || (props.workspaceList.find(ws => ws.isOpenInSidebar) || {id: undefined}).id
+    const idWsToOpen = idOpenInSidebar || props.currentWorkspace.id || undefined
 
     const fetchGetWorkspaceList = await props.dispatch(getMyselfWorkspaceList())
 
@@ -130,6 +136,15 @@ class Tracim extends React.Component {
       return true
     }
     return false
+  }
+
+  setDefaultUserLang = async (loggedUser) => {
+    const { props } = this
+    const fetchPutUserLang = await props.dispatch(putUserLang(loggedUser, props.user.lang))
+    switch (fetchPutUserLang.status) {
+      case 200: break
+      default: props.dispatch(newFlashMessage(props.t('Error while saving your default language')))
+    }
   }
 
   handleRemoveFlashMessage = msg => this.props.dispatch(removeFlashMessage(msg))
@@ -174,7 +189,7 @@ class Tracim extends React.Component {
           <Route path='/ui/workspaces/:idws?' render={() => [// Workspace Router
             // @FIXME - CH - 2018-03-26 - the use of array in a render function avoid having to wrap everything into
             // a wrapper div.
-            // This is required here to avoid having the div.tracim__content in the calendars pages.
+            // This is required here to avoid having the div.tracim__content in the agendas pages.
             // To fix this, upgrade React to at least 16.2.0 and use the first class component React.Fragment instead
             // of the array syntax that is kind of misleading. Also remove the key props
             <Route exact path={PAGE.WORKSPACE.ROOT} key='workspace_root' render={() =>
@@ -185,10 +200,10 @@ class Tracim extends React.Component {
               <Redirect to={{pathname: PAGE.WORKSPACE.CONTENT_LIST(props2.match.params.idws), state: {from: props.location}}} />
             } />,
 
-            <Switch>
+            <Switch key='workspace_content'>
               <Route
                 path={PAGE.WORKSPACE.CONTENT(':idws', ':type', ':idcts')}
-                key='workspace_content'
+                key='workspace_contentopen'
                 render={() =>
                   <div className='tracim__content fullWidthFullHeight'>
                     <WorkspaceContent />
@@ -212,7 +227,7 @@ class Tracim extends React.Component {
               </div>
             } />,
 
-            <Route path={PAGE.WORKSPACE.CALENDAR(':idws')} key='workspace_calendar' render={() =>
+            <Route path={PAGE.WORKSPACE.AGENDA(':idws')} key='workspace_agenda' render={() =>
               <AppFullscreenRouter />
             } />
           ]} />
@@ -224,7 +239,7 @@ class Tracim extends React.Component {
           <Route exact path={[
             PAGE.ADMIN.USER,
             PAGE.ADMIN.WORKSPACE,
-            PAGE.CALENDAR
+            PAGE.AGENDA
           ]} render={() => <AppFullscreenRouter />} />
 
           <Route path={'/wip/:cp'} component={WIPcomponent} /> {/* for testing purpose only */}
@@ -244,5 +259,5 @@ class Tracim extends React.Component {
   }
 }
 
-const mapStateToProps = ({ user, appList, contentType, workspaceList, flashMessage, system }) => ({ user, appList, contentType, workspaceList, flashMessage, system })
+const mapStateToProps = ({ user, appList, contentType, currentWorkspace, workspaceList, flashMessage, system }) => ({ user, appList, contentType, currentWorkspace, workspaceList, flashMessage, system })
 export default withRouter(connect(mapStateToProps)(translate()(Tracim)))