+++ functionnal_tests/cypress/integration/app_html-document/tiny_mce__can_type_content_spec.js
@@ -14,79 +14,78 @@ context('Known users as a workspace-manager', function () {
     cy.log('TODO find better way to deal with tinyMCE than using wait')
   })
 
-  it('Type into tiny mce and save from dashboard', function () {
-    return true // @FIXME - CH - 2018-02-07 - this test randomly fails
-    // I disable it until fixed
-    // see: https://github.com/tracim/tracim/issues/1327
 
+  it.only('Check TinyMCE is active if document is empty when opening', function () {
+    cy.get('[data-cy=dropdownCreateBtn]').first().click()
+    cy.contains('Write a document').click()
+    cy.get('[data-cy=createcontent__form__input]').type(DOCUMENT_TITLE)
+    cy.get('[data-cy=popup__createcontent__form__button]').click()
+    cy.get('.wsContentGeneric__header__close').click()
+    cy.contains(DOCUMENT_TITLE).click()
+
+    cy.waitForTinyMCELoaded().then(() => {
+      cy.assertTinyMCEIsActive()
+    })
+  })
+
+  it('Type into tiny mce and save from dashboard', function () {
     cy.visit(`/ui/workspaces/${this.workspace.workspace_id}/dashboard`)
-    cy.contains('Shared space manager')
     cy.get('[data-cy="contentTypeBtn_contents/html-document"]').click()
     cy.get('[data-cy=createcontent__form__input]').type(DOCUMENT_TITLE)
     cy.get('[data-cy=popup__createcontent__form__button]').click()
-    cy.wait(2000)
-    cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT)
-    cy.get('[data-cy=editionmode__button__submit]').click()
-    cy.contains(DOCUMENT_RAW_CONTENT).should('exist')
-    cy.wait(2000)
+
+    cy.waitForTinyMCELoaded()
+      .then(() => cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT))
+      .then(() => {
+        cy.get('[data-cy=editionmode__button__submit]').should('not.be.disabled').click()
+        cy.contains(DOCUMENT_RAW_CONTENT).should('exist')
+      })
   })
 
   it('Type into tiny mce and save', function () {
-    return true // @FIXME - CH - 2019-04-05 - this test randomly fails
-    // I disable it until fixed
-    // see: https://github.com/tracim/tracim/issues/1327
-
     cy.get('[data-cy=dropdownCreateBtn]').first().click()
     cy.contains('Write a document').click()
     cy.get('[data-cy=createcontent__form__input]').type(DOCUMENT_TITLE)
     cy.get('[data-cy=popup__createcontent__form__button]').click()
-    cy.wait(2000)
-    cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT)
-    cy.get('[data-cy=editionmode__button__submit]').click()
-    cy.contains(DOCUMENT_RAW_CONTENT).should('exist')
-    cy.wait(2000)
+
+    cy.waitForTinyMCELoaded()
+      .then(() => cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT))
+      .then(() => {
+        cy.get('[data-cy=editionmode__button__submit]').should('not.be.disabled').click()
+        cy.contains(DOCUMENT_RAW_CONTENT).should('exist')
+      })
   })
 
   it('Check if TinyMCE open content of the document', function () {
-    return true // @FIXME - CH - 2018-02-07 - this test randomly fails
-    // I disable it until fixed
-    // see: https://github.com/tracim/tracim/issues/1327
-
     cy.get('[data-cy=dropdownCreateBtn]').first().click()
     cy.contains('Write a document').click()
     cy.get('[data-cy=createcontent__form__input]').type(DOCUMENT_TITLE)
     cy.get('[data-cy=popup__createcontent__form__button]').click()
-    cy.wait(2000)
-    cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT)
-    cy.get('[data-cy=editionmode__button__submit]').click()
-    cy.get('[data-cy=wsContentGeneric__option__menu__addversion]').should('not.be.disabled').click()
-    cy.assertTinyMCEContent(DOCUMENT_HTML_CONTENT)
+
+    cy.waitForTinyMCELoaded()
+      .then(() => cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT))
+      .then(() => {
+        cy.get('[data-cy=editionmode__button__submit]').should('not.be.disabled').click()
+        cy.get('[data-cy=wsContentGeneric__option__menu__addversion]').should('not.be.disabled').click()
+        cy.assertTinyMCEContent(DOCUMENT_HTML_CONTENT)
+      })
   })
 
-  it('Check TinyMCE is active if document is empty when opening', function () {
-    cy.get('[data-cy=dropdownCreateBtn]').first().click()
-    cy.contains('Write a document').click()
-    cy.get('[data-cy=createcontent__form__input]').type(DOCUMENT_TITLE)
-    cy.get('[data-cy=popup__createcontent__form__button]').click()
-    cy.wait(2000)
-    cy.get('.wsContentGeneric__header__close').click()
-    cy.contains(DOCUMENT_TITLE).click()
-    cy.wait(2000)
-    cy.assertTinyMCEIsActive()
-  })
 
   it('Check TinyMCE is not active when opening if document has been modified', function () {
     cy.get('[data-cy=dropdownCreateBtn]').first().click()
     cy.contains('Write a document').click()
     cy.get('[data-cy=createcontent__form__input]').type(DOCUMENT_TITLE)
     cy.get('[data-cy=popup__createcontent__form__button]').click()
-    cy.wait(2000)
-    cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT)
-    cy.get('[data-cy=editionmode__button__submit]').should('not.be.disabled').click()
-    cy.wait(2000)
-    cy.get('.wsContentGeneric__header__close').click()
-    cy.contains(DOCUMENT_TITLE).click()
-    cy.wait(2000)
-    cy.assertTinyMCEIsActive(false)
+
+    cy.waitForTinyMCELoaded().then(() => {
+      cy.typeInTinyMCE(DOCUMENT_HTML_CONTENT)
+      cy.get('[data-cy=editionmode__button__submit]').should('not.be.disabled').click()
+      cy.get('[data-cy=popinFixed__header__button__close]').click()
+      cy.get('[data-cy=popinFixed]').should('not.exist')
+
+      cy.contains(DOCUMENT_TITLE).click()
+      cy.assertTinyMCEIsActive(false)
+    })
   })
 })